+++ templates/widget/template/demo-app/webpack.config.js
@@ -47,6 +47,18 @@ module.exports = {
         }, {
             test: /\.sht$/,
             loader: 'sham-ui-templates-loader?hot'
+        }, {
+            test: /\.sfw$/,
+            use: [
+                { loader: 'babel-loader' },
+                {
+                    loader: 'sham-ui-templates-loader?hot',
+                    options: {
+                        asModule: false,
+                        asSingleFileWidget: true
+                    }
+                }
+            ]
         } ]
     }
 };